# docs/api-contract.yaml
openapi: 3.0.3
info:
  title: Restart API
  version: 1.0.0
  description: |
    ## Overview
    Restart is a multi-tenant SaaS platform for managing instructor-led training programmes.
    Each instructor (tenant) operates in an isolated PostgreSQL schema — their data is
    completely separate from every other tenant.

    ## Two Scopes
    The API has two distinct actor scopes:

    - **ADMIN** — Internal staff. Data lives in the `public` schema. Manages tenant creation,
      instructor provisioning, tracks, weekly targets, and has read access across all tenants.
    - **INSTRUCTOR** — One per tenant. Data lives in the tenant's own schema. Manages their
      own cohorts, learners, and attendance records.

    ## Authentication
    All protected endpoints require a `Bearer` JWT in the `Authorization` header.

    - Admin tokens are obtained from `POST /api/auth/admin/login`.
    - Instructor tokens are obtained from `POST /api/auth/login` with an `X-Tenant-ID` header
      identifying which tenant schema to authenticate against.

    ### Why X-Tenant-ID on login?
    PostgreSQL schema-per-tenant means "john" in `tenant_a` and "john" in `tenant_b` are
    completely different users. The `X-Tenant-ID` header tells the backend which schema to
    search in during authentication. After login, the tenant is embedded in the JWT itself,
    so subsequent requests do **not** need `X-Tenant-ID` — the token carries it.

    ## Standard Response Envelope
...
    ```json
    {
...
    }
...
servers:
...
security:
  - BearerAuth: []
...
paths:
...
  /api/auth/admin/login:
    post:
      tags: [Auth]
      summary: Admin login
      description: |
        Authenticates an admin user against the `public.admin_users` table.
        Returns a JWT with `role: ADMIN` and no `tenantId` claim.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              username: admin
              password: secret123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseLoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Instructor login
      description: |
        Authenticates an instructor against the schema identified by `X-Tenant-ID`.
        Returns a JWT with `role: INSTRUCTOR` and a `tenantId` claim equal to the
        schema name. The frontend must store this token and send it on every subsequent
        instructor request — the `X-Tenant-ID` header is **not** needed after login.
      security: []
      parameters:
        - $ref: '#/components/parameters/XTenantID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              username: john_instructor
              password: myPassword
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseLoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ── TENANTS ──────────────────────────────────────────────────────────────
...
  /api/admin/tenants/{schemaName}/learners/bulk:
    post:
      tags: [Admin › Learners]
      summary: Bulk create learners in a tenant
      description: |
        Accepts an array of learner objects. Intended for CSV import flows where the
        client parses the CSV and sends the parsed array as JSON. All records are
        inserted in a single `saveAll` call within one transaction — if any record
        fails validation, none are saved.
      parameters:
        - $ref: '#/components/parameters/SchemaNamePath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/LearnerRequest'
      responses:
        '201':
          description: Learners created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseLearnerList'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/admin/tenants/{schemaName}/learners/{id}:
    get:
      tags: [Admin › Learners]
      summary: Get a learner by ID
      parameters:
        - $ref: '#/components/parameters/SchemaNamePath'
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Learner found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseLearnerResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Admin › Learners]
      summary: Update a learner
      parameters:
        - $ref: '#/components/parameters/SchemaNamePath'
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LearnerRequest'
      responses:
        '200':
          description: Learner updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseLearnerResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Admin › Learners]
      summary: Delete a learner
      parameters:
        - $ref: '#/components/parameters/SchemaNamePath'
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          $ref: '#/components/responses/EmptySuccess'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── ADMIN › ATTENDANCE ───────────────────────────────────────────────────
...
  /api/instructor/learners:
    post:
      tags: [Instructor › Learners]
      summary: Add a single learner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LearnerRequest'
      responses:
        '201':
          description: Learner created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseLearnerResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
    get:
      tags: [Instructor › Learners]
      summary: List all learners
      responses:
        '200':
          description: Learner list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseLearnerList'
  /api/instructor/learners/bulk:
    post:
      tags: [Instructor › Learners]
      summary: Bulk add learners
      description: |
        The client should parse the CSV on the frontend and post the resulting array.
        All rows are inserted atomically — a single invalid row rolls back everything.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/LearnerRequest'
      responses:
        '201':
          description: Learners created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseLearnerList'
  /api/instructor/learners/{id}:
    get:
      tags: [Instructor › Learners]
      summary: Get a learner by ID
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Learner found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseLearnerResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Instructor › Learners]
      summary: Update a learner
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LearnerRequest'
      responses:
        '200':
          description: Learner updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseLearnerResponse'
    delete:
      tags: [Instructor › Learners]
      summary: Delete a learner
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          $ref: '#/components/responses/EmptySuccess'

  # ── INSTRUCTOR › ATTENDANCE ──────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT issued by `/api/auth/admin/login` or `/api/auth/login`.
        Claims: `sub` (username), `role` (ADMIN | INSTRUCTOR), `tenantId` (schema name, null for admin).
        Expiry: 24 hours.
  parameters:
    SchemaNamePath:
      name: schemaName
      in: path
      required: true
      description: |
        The PostgreSQL schema name for the tenant (e.g. `kwame_mensah`).
        Derived from the instructor's name via `SchemaUtil.toSchemaName()`:
        lowercase, non-alphanumeric chars → underscore, collapse multiple underscores.
      schema:
        type: string
        example: kwame_mensah
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    XTenantID:
      name: X-Tenant-ID
      in: header
      required: true
      description: |
        The tenant schema name. Required **only** on `POST /api/auth/login`.
        All other instructor requests derive the tenant from the JWT `tenantId` claim.
      schema:
        type: string
        example: kwame_mensah

  # ── REQUEST BODIES ───────────────────────────────────────────────────────
  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: admin
        password:
          type: string
          format: password
          example: secret123
...
    LearnerRequest:
      type: object
      required: [fullname, email, phone]
      properties:
        fullname:
          type: string
          example: Kofi Mensah
        email:
          type: string
          format: email
          example: kofi.mensah@university.edu.gh
        phone:
          type: string
          pattern: '^0[2-5]\d{8}$'
          description: 10-digit Ghanaian mobile number starting with 02–05.
          example: "0241234567"
        gender:
          type: string
          nullable: true
          example: Male
        location:
          type: string
          nullable: true
          example: Accra
        region:
          type: string
          nullable: true
          example: Greater Accra
        institution:
          type: string
          nullable: true
          description: Educational institution or organization.
          example: University of Ghana
        graduated:
          type: boolean
          default: false
        cohortId:
          type: integer
          format: int64
          description: Optional cohort assignment.
          example: 1
...
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT Bearer token. Store in an httpOnly cookie on the frontend.
          example: eyJhbGciOiJIUzI1NiJ9...
...
    LearnerResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        fullname: { type: string }
        email: { type: string, format: email }
        phone: { type: string }
        gender: { type: string, nullable: true }
        location: { type: string, nullable: true }
        region: { type: string, nullable: true }
        institution: { type: string, nullable: true }
        graduated: { type: boolean }
        cohortId: { type: integer, format: int64, nullable: true }
        createdAt: { type: string, format: date-time }
...
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string }
        data: { nullable: true, example: null }

    # ── ENVELOPE WRAPPERS ─────────────────────────────────────────────────
    ApiResponseLoginResponse:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/LoginResponse'
...
    ApiResponseLearnerResponse:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/LearnerResponse'
    ApiResponseLearnerList:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/LearnerResponse'
...
    ApiEnvelope:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
        data: {}
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Full authentication is required to access this resource
            data: null
    BadRequest:
      description: Validation error or bad input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "phone: Phone must be a valid 10-digit Ghanaian number"
            data: null
    NotFound:
      description: Resource or tenant not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Tenant not found: unknown_schema"
            data: null
    EmptySuccess:
      description: Action completed, no data returned
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiEnvelope'
          example:
            success: true
            message: Success
            data: null